Commit 06/11/2022

+++ 01_SQL Basics/Uebungen/TeamsDb/Program.cs
@@ -2,28 +2,35 @@
 using Bogus;
 using Bogus.Extensions;
 using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
+using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
 
-Randomizer.Seed = new Random(1624);
-
-var opt = new DbContextOptionsBuilder()
-    .UseSqlite("Data Source=Teams.db")
-    .Options;
-
-using var db = new FilialContext(opt);
+if (args.Length < 1)
+{
+    Console.Error.WriteLine("Missing args.");
+    Console.Error.WriteLine("Usage: dotnet run -- (sqlserver|oracle|sqlite)");
+    return;
+}
+var options = MultiDbContext.GetConnectionInteractive(dbms: args[0].ToLower(), database: "TeamsDb");
+if (options is null) { return; }
+using var db = new TeamsContext(options);
 db.Database.EnsureDeleted();
 db.Database.EnsureCreated();
 
+Randomizer.Seed = new Random(1624);
 var departments = new string[] { "KIF", "AIF" };
 var subjects = new string[] { "POS", "DBI", "AM" };
 
-
+int rownr = 1;
+int rownr2 = 1;
 var teams = new Faker<Team>("de").CustomInstantiator(f =>
 {
     var klasse = $"{f.Random.Int(0, 2) * 2 + 1}{f.Random.String2(1, "ABC")}{f.Random.ListItem(departments)}";
     return new Team
     {
+        Id = rownr++,
         Name = $"SJ21/22_{klasse}",
         Klasse = klasse,
         Students = new Faker<Student>("de").CustomInstantiator(f =>
@@ -31,6 +38,7 @@ var teams = new Faker<Team>("de").CustomInstantiator(f =>
             var gender = f.Random.Enum<Gender>();
             return new Student
             {
+                Id = rownr2++,
                 Vorname = f.Name.FirstName(gender == Gender.M ? Bogus.DataSets.Name.Gender.Male : Bogus.DataSets.Name.Gender.Female),
                 Nachname = f.Name.LastName(gender == Gender.M ? Bogus.DataSets.Name.Gender.Male : Bogus.DataSets.Name.Gender.Female),
                 Gender = gender
@@ -45,11 +53,12 @@ var teams = new Faker<Team>("de").CustomInstantiator(f =>
 db.Teams.AddRange(teams);
 db.SaveChanges();
 
-
+rownr = 1;
 var aufgaben = new Faker<Aufgabe>("de").CustomInstantiator(f =>
 {
     return new Aufgabe
     {
+        Id = rownr++,
         Subject = f.Random.ListItem(subjects),
         Name = f.Commerce.ProductDescription(),
         Team = f.Random.ListItem(teams),
@@ -82,16 +91,18 @@ db.Abgaben.AddRange(abgaben);
 db.SaveChanges();
 
 
-public class FilialContext : DbContext
+public class TeamsContext : MultiDbContext
 {
-    public FilialContext(DbContextOptions opt) : base(opt)
+    public TeamsContext(DbContextOptions opt) : base(opt)
     {
 
     }
 
     protected override void OnModelCreating(ModelBuilder mb)
     {
+        base.OnModelCreating(mb);
         mb.Entity<Abgabe>().HasKey(a => new { a.AufgabeId, a.StudentId });
+        mb.Entity<Abgabe>().HasOne(a => a.Student).WithMany().OnDelete(DeleteBehavior.Restrict);
     }
 
     public DbSet<Team> Teams => Set<Team>();
@@ -105,10 +116,11 @@ public class FilialContext : DbContext
 
 public class Team
 {
+    [DatabaseGenerated(DatabaseGeneratedOption.None)]
     public int Id { get; set; }
-    public string Name { get; set; }
-    public string Klasse { get; set; }
-    public List<Student> Students { get; set; }
+    public string Name { get; set; } = default!;
+    public string Klasse { get; set; } = default!;
+    public List<Student> Students { get; set; } = default!;
 }
 
 
@@ -117,11 +129,12 @@ public class Team
 
 public class Aufgabe
 {
+    [DatabaseGenerated(DatabaseGeneratedOption.None)]
     public int Id { get; set; }
     public int TeamId { get; set; }
-    public string Subject { get; set; }
-    public string Name { get; set; }
-    public Team Team { get; set; }
+    public string Subject { get; set; } = default!;
+    public string Name { get; set; } = default!;
+    public Team Team { get; set; } = default!;
     public int MaxPunkte { get; set; }
 
 }
@@ -130,9 +143,9 @@ public class Aufgabe
 public class Abgabe
 {
     public int AufgabeId { get; set; }
-    public Aufgabe Aufgabe { get; set; }
+    public Aufgabe Aufgabe { get; set; } = default!;
     public int StudentId { get; set; }
-    public Student Student { get; set; }
+    public Student Student { get; set; } = default!;
     public DateTime? KorrigiertAm { get; set; }
     public int? Punkte { get; set; }
 }
@@ -141,12 +154,13 @@ public class Abgabe
 
 public class Student
 {
+    [DatabaseGenerated(DatabaseGeneratedOption.None)]
     public int Id { get; set; }
-    public string Vorname { get; set; }
-    public string Nachname { get; set; }
+    public string Vorname { get; set; } = default!;
+    public string Nachname { get; set; } = default!;
     public Gender Gender { get; set; }
     public int TeamId { get; set; }
-    public Team Team { get; set; }
+    public Team Team { get; set; } = default!;
 }
 
 public enum Gender { F = 1, M }
